Stopped using the hungarian notation and DWORDs

+++ src/Games/AlphaMW2/GameFunctions.h
@@ -22,7 +22,7 @@ extern const char *(*Dvar_GetString)(const char *dvarName);
 
 extern playerState_s *(*GetPlayerState)(int clientNum);
 
-extern bool (*Session_IsHost)(DWORD sessionDataPtr, int clientNum);
+extern bool (*Session_IsHost)(uintptr_t sessionDataPtr, int clientNum);
 
 extern void (*SP_script_model)(gentity_s *mSelf);
